+++ Wordplay/proj/view/Properties/AssemblyInfo.cs
